Added bash script to run all example scripts.
This is in an effort to confirm that all example scripts are up to date and
that their behavior has not changed. In doing so, an error has already
been caught in integer rounding that is no longer necessary.

+++ docs/tutorials/scripts/dmtools/clamping_ex.py
@@ -9,11 +9,11 @@ def f(x):
 image = dmtools.read_png('checks_10.png')
 scaled_image = transform.rescale(image, k=10, weighting_function=f, support=7)
 
-clip_image = transform.clip(scaled_image).astype(np.uint8)
+clip_image = transform.clip(scaled_image)
 dmtools.write_png(clip_image, 'checks_10_clip.png')
 
-normalize_image = transform.normalize(scaled_image).astype(np.uint8)
+normalize_image = transform.normalize(scaled_image)
 dmtools.write_png(normalize_image, 'checks_10_normalize.png')
 
-wraparound_image = transform.wraparound(scaled_image).astype(np.uint8)
+wraparound_image = transform.wraparound(scaled_image)
 dmtools.write_png(wraparound_image, 'checks_10_wraparound.png')
